Revert "COZMO-15195: Feeding Cubes draining too quickly"
This reverts commit 84d77308b31d488b88f6061c14ad0875191530e2.

+++ engine/behaviorSystem/activities/activities/activityFeeding.h
@@ -42,7 +42,7 @@ public:
   virtual Result Update(Robot& robot) override;
   
   // Implementation of IFeedingListener
-  virtual void StartedEating(Robot& robot, const float duration_s) override;
+  virtual void StartedEating(Robot& robot, const int duration_s) override;
   virtual void EatingComplete(Robot& robot) override;
   virtual void EatingInterrupted(Robot& robot) override;
